Moved functions for raven, blender, and utils to separate files, updated setup.py
